+++ yggdrasil/examples/gs_lesson4b/gs_lesson4b.sh
@@ -27,9 +27,13 @@ case $1 in
 	echo "Running R"
 	yaml='gs_lesson4b_r.yml'
 	;;
-    -f | --fortran)
+    -f | --fortran )
 	echo "Running Fortran"
 	yaml='gs_lesson4b_fortran.yml'
+	;;
+    -j | --julia)
+	echo "Running Julia"
+	yaml='gs_lesson4b_julia.yml'
 	;;
 esac